Move src/app/client/src/components/ into src/app/client/src/components/submission/ and fix sizing of submission layout
