+++ src/app/client/src/components/submission/NavBar.tsx
